+++ resources/skins.femiwiki.mobile/init.js
@@ -1,24 +1,35 @@
-var newWikitext = mw.user.options.get('visualeditor-newwikitext') === 1;
+// Note: A preference of MediaWiki is stored as string always.
+var newWikitext = mw.user.options.get('visualeditor-newwikitext') === '1';
 
 /**
+ * @param {Event} ev
  * @return {void}
  */
-function click(/**@type Event*/ e) {
-  // @ts-ignore
-  var href = this.getAttribute('href') || e.target.toString();
-  var veaction = /\bveaction\b/.test(href);
-  var supported = !veaction && !newWikitext;
-
-  if (supported) {
-    // Do noting here to allow the user to move to href.
+function click(ev) {
+  if (ev === null || ev.currentTarget === null) {
     return;
   }
-  e.preventDefault();
-  var switchUri = new mw.Uri(href).extend({
+  var anchor = /** @type {HTMLAnchorElement} */ (ev.currentTarget);
+  var url = new mw.Uri(anchor.href);
+  var visualEditing =
+    // VisualEdit
+    url.query.veaction === 'edit' ||
+    // If the new wikitext mode is enabled, all cases of edit will be done on VisualEditor
+    newWikitext;
+
+  if (!visualEditing) {
+    // Do nothing here to allow the user to move to href.
+    return;
+  }
+  ev.preventDefault();
+  var switchUri = url.extend({
     mobileaction: 'toggle_view_desktop',
   });
 
   mw.loader.using(['oojs', 'oojs-ui-windows'], function () {
+    if (!OO.ui) {
+      return;
+    }
     OO.ui
       .confirm(mw.msg('skin-femiwiki-desktop-switch-confirm'))
       .done(function (confirmed) {
@@ -54,7 +65,7 @@ function main() {
 
   /** @type NodeListOf<Element> */ var $allEditLinks =
     document.querySelectorAll(
-      '#ca-ve-edit, #ca-edit, .mw-editsection a, .edit-link'
+      '#ca-ve-edit a, #ca-edit a, .mw-editsection a, .edit-link'
     );
 
   for (var i = 0; i < $allEditLinks.length; i++) {