+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 - The styles for plain links are disabled by default now. To recover the previous behavior, add `$wgFemiwikiAddLinkClass = true;` to your LocalSettings.php. (https://github.com/femiwiki/FemiwikiSkin/pull/469)
 - The elements on the skin are now larger than before by default. Users can disable this on Special:Preferences.
+- Fixed the bug that the desktop change confirming dialog is not affected by the wikitext mode in VisualEditor preference on Mobile.
+- Fixed the bug of the desktop change confirming dialog on mobile.
 
 ## v1.10.10